feat : criado classe ManipularProducao, todos os métodos relacionados ao módulo 4 estão contidas nela e a MainModulo4 apenas faz a chamada. feat: criado classe Extra com as funções de validar int e float
